requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy<2.0
scipy>=1.7.0
matplotlib>=3.5.0
scikit-image>=0.19.0
Pillow>=9.0.0
tqdm>=4.65.0
opencv-python>=4.7.0
imageio>=2.25.0